+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Unreleased
 
 - utils.grep: Fix wrong number of arguments.
 - utils.diff: Fix comparison between different bytes types.
+- lib.platforms.director: Add IMapV4, fix ScriptCastV4.
 
 0.7.2 - 2019-11-24
 ==================